# doc/apidoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Khazana POS API",
    "description": "## Khazana POS Backend API\n\nAvailable modules:\n- **Auth** – Login, profile, change password\n- **Users** – User management and role assignment\n- **Sales** – Cash, credit, VAT cash, VAT credit sales\n- **Inventory** – Items, stock receive / issue / adjust\n- **Customers** – Customer management, ledger, payments\n- **Roles** – Role management\n- **Menus** – Menu items\n- **Permissions** – Role-based permissions\n- **Orders** – Sales orders (regular and VAT)\n- **Pricing** – Sale prices and cost prices\n- **Finance** – Money receive and cash purchases\n- **Packets** – Packet stock management\n- **NC Adjustment** – Non-conformance adjustments\n- **Assortment** – Assortment records\n- **Reports** – Sales, stock, item-wise, customer statements\n- **Admin** – Branches, settings, banks, audit log\n- **Dashboard** – Summary statistics\n- **Categories** – Item category management",
    "version": "1.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "required": ["userName", "password"],
        "properties": {
          "userName": { "type": "string", "example": "john_doe" },
          "password": { "type": "string", "example": "password123" },
          "name": { "type": "string", "example": "John Doe" },
          "branchId": { "type": "integer", "example": 1 }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 },
          "hasNextPage": { "type": "boolean", "example": true },
          "hasPreviousPage": { "type": "boolean", "example": false }
        }
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users",
        "operationId": "UsersController_findAll",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 1, "minimum": 1 }, "description": "Page number" },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 10, "minimum": 1, "maximum": 100 }, "description": "Items per page" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "User retrieved successfully" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": { "type": "array", "items": { "type": "object" } },
                        "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "operationId": "UsersController_create",
        "security": [{ "access-token": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDto" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "409": { "description": "Username already exists" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "operationId": "UsersController_findOne",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "User UUID" }
        ],
        "responses": {
          "200": { "description": "User found" },
          "404": { "description": "User not found" }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update user by ID",
        "operationId": "UsersController_update",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "User UUID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDto" }
            }
          }
        },
        "responses": {
          "200": { "description": "User updated successfully" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user by ID",
        "operationId": "UsersController_remove",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "User UUID" }
        ],
        "responses": {
          "200": { "description": "User deleted successfully" }
        }
      }
    },
    "/users/{id}/roles": {
      "post": {
        "tags": ["Users"],
        "summary": "Set roles for a user",
        "operationId": "UsersController_setRoles",
        "security": [{ "access-token": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "User UUID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "roleId": { "type": "string", "format": "uuid" },
                        "branchId": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Roles assigned successfully" }
        }
      }
    },
...
  },
...
}